test: repair `multi-instance` tests
